css/css-regions/parsing/flow-from-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Regions Module Level 1: parsing flow-from with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-regions/#propdef-flow-from">
<meta name="assert" content="flow-from supports only the grammar '<ident> | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("flow-from", "0");
test_invalid_value("flow-from", '"string"');
test_invalid_value("flow-from", "none COLUMN");
test_invalid_value("flow-from", "COLUMN none");
</script>
</body>
</html>
